refactor: viewerの共通コンポーネント抽出によるコード量削減 (#109)
- LoadingSpinner共通コンポーネント作成（6ファイルの重複解消）
- EmptyState共通コンポーネント作成（5ファイルの重複解消）
- MessageToast共通コンポーネント作成（DatasetManager, QueueManagerの重複解消）
- TypeBadge共通コンポーネント作成（FavoritePage, DatasetManagerの重複解消）
- FormControls共通コンポーネント作成（ConfigEditorから8つのUI部品を抽出）
- アニメーションバリアント共通化（AppView, Sidebarの重複解消）

Closes #109
