+++ viewer/src/components/MemoTab.tsx
@@ -1,7 +1,7 @@
-import { motion } from "motion/react";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { fetchMemo, saveMemo } from "../api";
 import { MarkdownPane } from "./MarkdownPane";
+import { LoadingSpinner } from "./shared/LoadingSpinner";
 
 type MemoView = "edit" | "preview";
 
@@ -42,23 +42,7 @@ export function MemoTab({
   }, [appName, content, hasChanges, saving]);
 
   if (loading) {
-    return (
-      <div className="flex h-full items-center justify-center">
-        <motion.div
-          className="flex flex-col items-center gap-3"
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          transition={{ duration: 0.3 }}
-        >
-          <motion.div
-            className="size-8 rounded-full border-2 border-indigo-800 border-t-indigo-400"
-            animate={{ rotate: 360 }}
-            transition={{ duration: 1, repeat: Number.POSITIVE_INFINITY, ease: "linear" }}
-          />
-          <p className="text-xs text-gray-500">読み込み中...</p>
-        </motion.div>
-      </div>
-    );
+    return <LoadingSpinner />;
   }
 
   /* ── Mobile layout ── */
